add test cover full types

+++ flink-cdc-connect/flink-cdc-pipeline-connectors/flink-cdc-pipeline-connector-hudi/src/main/java/org/apache/flink/cdc/connectors/hudi/sink/util/RowDataUtils.java
@@ -146,7 +146,7 @@ public class RowDataUtils {
                 fieldGetter = row -> row.getByte(fieldPos);
                 break;
             case SMALLINT:
-                fieldGetter = row -> row.getInt(fieldPos);
+                fieldGetter = row -> row.getShort(fieldPos);
                 break;
             case BIGINT:
                 fieldGetter = row -> row.getLong(fieldPos);
